size_t instead of unsigned int for kmalloc

+++ src/kernel/include/kernel/stdlib.h
@@ -1,8 +1,9 @@
 #ifndef _KERNEL_STDLIB_H
 #define _KERNEL_STDLIB_H
 
+#include <stddef.h>
 void init_kmalloc(void);
-void *kmalloc(unsigned int size);
+void *kmalloc(size_t size);
 void kfree(void *ptr);
 
 #endif // _KERNEL_STDLIB_H
